requirements.txt
altgraph==0.17.3
cryptocode==0.1
future==0.18.2
pefile==2022.5.30
Pillow==9.3.0
pycryptodomex==3.16.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pywin32-ctypes==0.2.0